Create Tracing_Game.ino
This is the code for a Memory game that Yishu and I made. The code is made to run on a Teensiduino Micro-controller attached to the computer by USB. This was made for the Electronics Final Project in PHYS 247 at St. Olaf College. The circuit itself is able to be used in many different applications.