Create Clock.ino
This is a digital clock programmed to accurately display the time up to the most recent 5-minute mark. The clock is as accurate as you set it, but for lack of display room it only shows the most recent 5 minute time on the outside of its display, moving like a minute hand. It needs to be set with the Unix epoch time. It makes use of the Time library in Arduino. The circuit used is linked in the code and can be used for the Tracing Game program I also wrote.